refactor: reorganize project structure by moving core modules into dedicated comms, control, drivers, runtime, safety, and tasks subdirectories
