+++ edge/app/comms/crc16.cpp
