+++ edge/app/comms/crc16.hpp
