+++ edge/app/comms/hmac_sha256.hpp
