+++ edge/app/comms/i_serial.hpp
