+++ edge/app/comms/telemetry.hpp
