+++ edge/app/control/domain.hpp
