+++ edge/app/control/state_machine.cpp
