+++ edge/app/control/state_machine.hpp
