+++ edge/app/drivers/platform_io.cpp
