+++ edge/app/drivers/platform_io.hpp
