+++ edge/app/safety/fail_safe_supervisor.hpp
