+++ edge/app/safety/freertos_hooks.cpp
