+++ edge/app/safety/panic.hpp
