+++ edge/app/safety/watchdog.hpp
